requirements.txt
duckdb
pydantic>=2.7
pyyaml
typer[all]
rich
requests
jinja2
reportlab
